actor & critic update
