+++ rl/algorithms/sac.py
@@ -1,6 +1,6 @@
 import torch
 from catalyst.dl.utils import UtilsFactory
-import catalyst.rl.networks.agents as agents
+import catalyst.rl.agents.agents as agents
 from catalyst.rl.algorithms.base import BaseAlgorithm, soft_update
